Finished StatsCommand for Flocking and PD
