+++ src/Flocking/FlockingFactory.java
@@ -1,0 +1,16 @@
+package Flocking;
+
+import Simstation.*;
+import mvc.*;
+
+class FlockingFactory extends SimulationFactory {
+    @Override
+    public Model makeModel() {
+        return new FlockingSimulation();
+    }
+
+    @Override
+    public String getTitle() {
+        return "Flocking Simulation";
+    }
+}
