+++ src/Simstation/StatsCommand.java
@@ -7,8 +7,7 @@ import Plague.*;
 import mvc.*;
 
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 public class StatsCommand extends Command {
     public StatsCommand(Model model) {
@@ -29,10 +28,28 @@ public class StatsCommand extends Command {
             double infectedPercent = plagueSimulation.getPercentInfected();
             stats.add("%infected = " + infectedPercent);
         } else if (model instanceof FlockingSimulation) {
+            int[] speedsCount = new int[5];  // Array to hold counts for speeds 1-5
 
+            Iterator<Agent> it = simulation.agentIterator();
+            while (it.hasNext()) {
+                Agent agent = it.next();
+                if (agent instanceof Bird) {
+                    Bird bird = (Bird) agent;
+                    int speed = bird.getSpeed();  // Assume Bird has a getSpeed() method
+                    if (speed >= 1 && speed <= 5) {
+                        speedsCount[speed - 1]++;
+                    }
+                }
+            }
+
+            for (int i = 0; i < speedsCount.length; i++) {
+                stats.add("#birds @ speed " + (i + 1) + " = " + speedsCount[i]);
+            }
         }
         else if (model instanceof PrisonersDilemmaSimulation) {
-
+            PrisonersDilemmaSimulation pdSim = (PrisonersDilemmaSimulation) model;
+            stats.add("# Prisoners: " + pdSim.AGENT_POPULATION);
+            stats.add("Average Fitness: " + pdSim.getAverageFitness());
         }
 
         return stats.toArray(new String[0]);
